Fix: :bug: timezone issue fixed

+++ app/Http/Controllers/InvoiceController.php
@@ -1011,8 +1011,8 @@ class InvoiceController extends Controller
             if (!empty($school->timezone)) {
                 $timeZone = $school->timezone;
             }
-            $data['p_billing_period_start_date'] = $this->formatDateTimeZone($data['p_billing_period_start_date'], 'long', $timeZone,'UTC',1);
-            $data['p_billing_period_end_date'] = $this->formatDateTimeZone($data['p_billing_period_end_date'], 'long',$timeZone,'UTC',1);
+            $data['p_billing_period_start_date'] = $this->formatDateTimeZone($data['p_billing_period_start_date'], 'short', $timeZone,'UTC',1);
+            $data['p_billing_period_end_date'] = $this->formatDateTimeZone($data['p_billing_period_end_date'], 'short',$timeZone,'UTC',1);
              
             $dateS = date('Y-m-d', strtotime(str_replace('.', '-', $data['p_billing_period_start_date'])));
             $dateEnd = date('Y-m-d', strtotime(str_replace('.', '-', $data['p_billing_period_end_date'])));
